Refactor Session to interface, rename implementation to SynchronizedSession
